Moving towards Signpost
We want to move towards Signpost as the current netflix one is a bit messy in terms of intrusion. With Signpost, you can use any http client in order to access the backend (apache). It is much cleaner and will help towards building the service. I also deleted the OAuth test which failed the instrumentation. Now you should be able to instrument easily by right-clicking on the test application and run it.