Moving away from netflix into signpost
There is no more need for netflix's lib anymore and we are using signpost. I am having verify errors in the manual test which I am not really sure how to work out for now. The signpost lib is much less intrusive and we can use the http commons to do the calls which is much nicer. This will come in handy when creating the service.